Add unit and integration tests for Item service
The commit adds unit tests in item_test.go for the "GetItemByID" and "GetAllItems" functions of the Item service. A new MockRepository structure is used to mock the results of these functions. Additionally, integration tests were implemented in item_integration_test.go for the same functions. These tests operate on an actual database connection, managed by the dbconn package.