Handle items with null category
This commit updates the handling of items with a null CategoryID. A fallback default category "Uncategorized" is added for these items to avoid null values. This is done in both itemDomain and get_items_with_category files. Also, the JOIN statements in SQL queries are changed to LEFT JOINs to account for items without a category. The data type of field CategoryName is changed from string to pgtype.Text for consistency.